# frontend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/Flights/{id}": {
      "get": {
        "tags": [
          "Flights"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FlightDetailResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightDetailResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/FlightsAdmin/weekly-schedule": {
      "post": {
        "tags": [
          "FlightsAdmin"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWeeklyScheduleDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWeeklyScheduleDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWeeklyScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlightManagementResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlightManagementResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlightManagementResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/admin/FlightsAdmin": {
      "post": {
        "tags": [
          "FlightsAdmin"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFlightDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFlightDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFlightDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FlightManagementResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightManagementResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightManagementResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "FlightsAdmin"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlightManagementResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlightManagementResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlightManagementResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/admin/FlightsAdmin/{flightId}": {
      "put": {
        "tags": [
          "FlightsAdmin"
        ],
        "parameters": [
          {
            "name": "flightId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFlightDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFlightDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFlightDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "FlightsAdmin"
        ],
        "parameters": [
          {
            "name": "flightId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/admin/FlightsAdmin/routes": {
      "post": {
        "tags": [
          "FlightsAdmin"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRouteDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRouteDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRouteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RouteManagementResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteManagementResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteManagementResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "FlightsAdmin"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RouteManagementResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RouteManagementResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RouteManagementResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/admin/FlightsAdmin/routes/{routeId}": {
      "put": {
        "tags": [
          "FlightsAdmin"
        ],
        "parameters": [
          {
            "name": "routeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRouteDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRouteDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRouteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateFlightDto": {
        "type": "object",
        "properties": {
          "flightNumber": {
            "type": "string",
            "nullable": true
          },
          "routeId": {
            "type": "integer",
            "format": "int32"
          },
          "aircraftId": {
            "type": "integer",
            "format": "int32"
          },
          "departureTime": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalTime": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "CreateRouteDto": {
        "type": "object",
        "properties": {
          "departureAirportId": {
            "type": "integer",
            "format": "int32"
          },
          "arrivalAirportId": {
            "type": "integer",
            "format": "int32"
          },
          "distanceKm": {
            "type": "integer",
            "format": "int32"
          },
          "estimatedDurationMinutes": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CreateWeeklyScheduleDto": {
        "type": "object",
        "properties": {
          "weekStartDate": {
            "type": "string",
            "format": "date"
          },
          "autoGenerateWeeks": {
            "type": "integer",
            "format": "int32"
          },
          "flights": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeeklyFlightPatternDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FlightDetailResponse": {
        "type": "object",
        "properties": {
          "flightId": {
            "type": "integer",
            "format": "int32"
          },
          "flightNumber": {
            "type": "string",
            "nullable": true
          },
          "departureTime": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalTime": {
            "type": "string",
            "format": "date-time"
          },
          "departureAirport": {
            "type": "string",
            "nullable": true
          },
          "arrivalAirport": {
            "type": "string",
            "nullable": true
          },
          "distanceKm": {
            "type": "integer",
            "format": "int32"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "aircraftModel": {
            "type": "string",
            "nullable": true
          },
          "seatInventory": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/SeatClassDetail"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FlightManagementResponse": {
        "type": "object",
        "properties": {
          "flightId": {
            "type": "integer",
            "format": "int32"
          },
          "flightNumber": {
            "type": "string",
            "nullable": true
          },
          "routeCode": {
            "type": "string",
            "nullable": true
          },
          "aircraftModel": {
            "type": "string",
            "nullable": true
          },
          "departureTime": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalTime": {
            "type": "string",
            "format": "date-time"
          },
          "totalSeats": {
            "type": "integer",
            "format": "int32"
          },
          "availableSeats": {
            "type": "integer",
            "format": "int32"
          },
          "bookedSeats": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "RouteManagementResponse": {
        "type": "object",
        "properties": {
          "routeId": {
            "type": "integer",
            "format": "int32"
          },
          "departureAirport": {
            "type": "string",
            "nullable": true
          },
          "arrivalAirport": {
            "type": "string",
            "nullable": true
          },
          "distanceKm": {
            "type": "integer",
            "format": "int32"
          },
          "estimatedDurationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "activeFlights": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "SeatClassDetail": {
        "type": "object",
        "properties": {
          "seatClassId": {
            "type": "integer",
            "format": "int32"
          },
          "className": {
            "type": "string",
            "nullable": true
          },
          "totalSeats": {
            "type": "integer",
            "format": "int32"
          },
          "availableSeats": {
            "type": "integer",
            "format": "int32"
          },
          "heldSeats": {
            "type": "integer",
            "format": "int32"
          },
          "soldSeats": {
            "type": "integer",
            "format": "int32"
          },
          "currentPrice": {
            "type": "number",
            "format": "double"
          },
          "basePrice": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateFlightDto": {
        "type": "object",
        "properties": {
          "flightNumber": {
            "type": "string",
            "nullable": true
          },
          "aircraftId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "departureTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "arrivalTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateRouteDto": {
        "type": "object",
        "properties": {
          "departureAirportId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "arrivalAirportId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "distanceKm": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "estimatedDurationMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WeeklyFlightPatternDto": {
        "type": "object",
        "properties": {
          "flightNumber": {
            "type": "string",
            "nullable": true
          },
          "flightNumberPrefix": {
            "type": "string",
            "nullable": true
          },
          "routeId": {
            "type": "integer",
            "format": "int32"
          },
          "aircraftId": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "type": "integer",
            "format": "int32"
          },
          "departureTimeOfDay": {
            "type": "string",
            "format": "time"
          },
          "arrivalTimeOfDay": {
            "type": "string",
            "format": "time"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter JWT Bearer token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}